LayoutTests/fast/forms/time/time-editable-components/time-editable-components-input-and-change-events-expected.txt

Test for input and change events for <input type=time>

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS inputEventsFired is 0
PASS changeEventsFired is 0
PASS inputEventsFired is 0
PASS changeEventsFired is 0
PASS inputEventsFired is 1
PASS changeEventsFired is 0
PASS input.value is "00:02"
PASS inputEventsFired is 1
PASS changeEventsFired is 1
PASS successfullyParsed is true

TEST COMPLETE
